resume_screening/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.9.1
arabic-reshaper==2.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
Aspose.Slides==22.10.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
authy==2.2.6
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
blis==1.2.0
catalogue==2.0.10
celery==5.4.0
certifi==2024.12.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cryptography==38.0.1
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.16
django-allauth==65.2.0
django-crispy-forms==1.14.0
django-excel-response==2.0.5
django-password-reset==2.0
djangorestframework==3.14.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.17.0
fiscalyear==0.4.0
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.12.0
future==0.18.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.28.1
idna==3.4
importlib-metadata==5.2.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
isoduration==20.11.0
jdcal==1.4.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.6.17
kiwisolver==1.4.7
kombu==5.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==4.9.1
marisa-trie==1.2.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==9.0.0
mpmath==1.3.0
murmurhash==1.0.12
mysql-connector==2.2.9
mysql-connector-python==9.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.23.4
openai==1.60.2
openpyxl==2.4.11
oscrypto==1.3.0
overrides==7.7.0
packaging==24.1
pandas==1.5.2
pandocfilters==1.5.1
parso==0.8.4
pdf2pptx==1.0.5
phonenumbers==8.13.3
Pillow==9.2.0
platformdirs==4.2.2
plotly==5.11.0
preshed==3.0.9
prettytable==3.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.21
pycryptodome==3.15.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
PyMuPDF==1.20.1
pyparsing==3.2.1
PyPDF2==3.0.1
PyPDF3==1.0.6
pypng==0.20220715.0
PyQRCode==1.2.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.2
python-json-logger==2.0.7
python-pptx==0.6.21
python-slugify==8.0.4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0
pyzmq==26.2.0
qrcode==6.1
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
reportlab==3.6.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.3
srsly==2.5.1
stack-data==0.6.3
svglib==1.4.1
sympy==1.13.1
tenacity==8.1.0
terminado==0.18.1
text-unidecode==1.3
thinc==8.3.4
threadpoolctl==3.5.0
tinycss2==1.1.1
tokenizers==0.21.0
tomli==2.0.1
torch==2.6.0
tornado==6.4.1
tqdm==4.64.1
traitlets==5.14.3
transformers==4.48.2
typer==0.15.1
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==4.2
uri-template==1.3.0
uritools==4.0.0
urllib3==1.26.12
vine==5.1.0
Wand==0.6.11
wasabi==1.1.3
wcwidth==0.2.5
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.17.2
xhtml2pdf==0.2.8
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.11.0
